pyproject.toml
[project]
name = "slurm-usage-history"
dynamic = ["version"]
authors = [
  { name = "Sören Wacker", email = "s.wacker@gmail.com" },
]
description = "A great package."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "pandas",
  "pyarrow",
  "dash",
  "tqdm",
  "dash-bootstrap-components",
  "python-dotenv",
  "flask-saml",
  "duckdb",
  "gunicorn",
  "tabulate",
  "dask[dataframe]",
  "pip>=24.3.1",
]

[tool.uv]
dev-dependencies = [
  "hatch>=1.13.0,<2",
  "pre-commit>=3.8.0,<4",
  "pytest>=8.3.3,<9",
  "pytest-cov>=5.0.0,<6",
  "mkdocstrings[python]>=0.26.1,<1",
  "mkdocs-material>=9.5.39,<10",
  "setuptools-scm>=8.1.0,<9",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "_version.py"

[project.urls]
Homepage = "https://gitlab.ewi.tudelft.nl/sdrwacker/slurm-usage-history"

[project.scripts]
slushi-waiting-times = "slurm_usage_history.scripts.waiting_times:main"
slushi-get-weekly-usage = "slurm_usage_history.scripts.get_weekly_usage:main"
slushi-dashboard = "slurm_usage_history.app.main:main"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]

[tool.coverage]
run.source = ["slurm_usage_history"]
port.exclude_lines = [
  'pragma: no cover',
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]



[tool.ruff]
src = ["src"]
exclude = []
line-length = 200  # how long you want lines to be, 88 is black's default

[tool.ruff.format]
docstring-code-format = true  # code snippets in docstrings will be formatted

[tool.ruff.lint]
# For more strict linting and code simplifications, replace this with
# https://github.com/NLeSC/python-template/blob/main/template/pyproject.toml.jinja#L121-L136
select = [
  "E", "F", "W", # flake8
  "B",           # flake8-bugbear
  "I",           # isort
  "ARG",         # flake8-unused-arguments
  "C4",          # flake8-comprehensions
  "EM",          # flake8-errmsg
  "ICN",         # flake8-import-conventions
  "ISC",         # flake8-implicit-str-concat
  "G",           # flake8-logging-format
  "PGH",         # pygrep-hooks
  "PIE",         # flake8-pie
  "PL",          # pylint
  "PT",          # flake8-pytest-style
  "RET",         # flake8-return
  "RUF",         # Ruff-specific
  "SIM",         # flake8-simplify
  "UP",          # pyupgrade
  "YTT",         # flake8-2020
  "EXE",         # flake8-executable
  # "D",           # pydocstyle, uncomment to have the docstrings linted
  # "ANN",         # flake8-annotations, uncomment to have type hint annotations linted
]
ignore = [
  "PLR",    # Design related pylint codes
  "ISC001", # Conflicts with formatter
]
pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
# Tests can ignore a few extra rules
"tests/**.py" = [
    "ANN201", # Missing return type annotation for public function
    "D103",   # Missing docstring
    "S101",   # Use of assert is detected
    "INP001", # Missing __init__.py file
]

[tool.yamlfix]
line_length = 88
sequence_style = "keep_style"
explicit_start = false
whitelines = 1
section_whitelines = 1
